Refactor optimizers from iterative_estimator into an optim_lib.
PiperOrigin-RevId: 349601983
